+++ CHANGELOG.md
@@ -1,3 +1,5 @@
+* NP-756 Advisernet related content component
+
 ## <sub>v1.9.3</sub>
 
 #### _Sep. 16, 2020_